Moved foreign declarations to PathfinderComposition.
